// silap-colmirwcl/src/app/modules/silapapi/model/formRicercaVersamentoEsoneri.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface FormRicercaVersamentoEsoneri { 
    annoProtocollo?: number;
    codFisc?: string;
    dataFine?: Date;
    dataInizio?: Date;
    denomAzienda?: string;
    flgDueRate?: string;
    flgUnicaSoluzione?: string;
    numProtocollo?: number;
    statoDichiarazione?: Array<number>;
}